Optimize syntax coloring.
bin/generate's runtime was dramatically increased by calls to
bin/highlight for server-side code coloring.  My guess was that
the overhead of loading all of the highlightjs modules in node
was slowing things down.

So, I added bin/highlightd, which does what bin/highlight did,
but instead of being a one-shot command line tool, it's a daemon
that reads commands (lines of JSON) from standard input, and writes
responses (lines of JSON) to standard output.  It's much faster.